// Ladol/resources/assets/js/types/evaluation.ts
export interface TEvaluation {
  id: number;
  bsc_measurement_period_id: number;
  user_id: number;
  department_id: number;
  performance_category_id: number;
  scorecard_score?: any;
  scorecard_self_score?: any;
  scorecard_percentage?: any;
  penalty_score?: any;
  behavioral_score?: any;
  behavioral_self_score?: any;
  behavioral_percentage?: any;
  weight_sum?: any;
  manager_id?: any;
  employee_strength?: any;
  employee_developmental_area?: any;
  special_achievement?: any;
  manager_approval_comment?: any;
  manager_approval_date?: any;
  manager_approval_approved?: any;
  manager_of_manager_id?: any;
  approval_status?: any;
  manager_of_manager_approved?: any;
  manager_of_manager_approval_comment?: any;
  manager_of_manager_approval_date?: any;
  kpi_submitted?: any;
  kpi_submitted_date?: any;
  kpi_accepted?: any;
  kpi_accepted_date?: any;
  appraisal_accepted?: any;
  appraisal_accepted_comment?: any;
  appraisal_accepted_date?: any;
  head_of_strategy_id?: any;
  head_of_hr_id?: any;
  head_of_hr_approved?: any;
  head_of_hr_approval_comment?: any;
  head_of_strategy_approval_comment?: any;
  head_of_strategy_approved?: any;
  head_of_hr_approved_date?: any;
  head_of_strategy_approved_date?: any;
  company_id: number;
  created_at: string;
  updated_at: string;
  is_disputed?: any;
  user: User;
  manager?: any;
  behavioral_evaluation_details: any[];
  department: Department;
  measurement_period: Measurementperiod;
}

interface Measurementperiod {
  id: number;
  from: string;
  to: string;
  created_at: string;
  updated_at: string;
  status: number;
  head_of_hr_id: number;
  head_of_strategy_id: number;
  scorecard_percentage: string;
  behavioral_percentage: string;
  type: string;
}

interface Department {
  id: number;
  name: string;
  company_id: number;
  manager_id: number;
  created_at: string;
  updated_at: string;
  color?: any;
  code: string;
}

interface User {
  id: number;
  emp_num: string;
  name: string;
  nin_no: string;
  bvn: string;
  tax_id: string;
  tax_authority: string;
  pension_id: string;
  pension_administrator: string;
  pension_type: string;
  medical_code?: any;
  alt_email: string;
  email: string;
  sex: string;
  dob: string;
  phone: string;
  alt_phone?: any;
  marital_status: string;
  image: string;
  address: string;
  hiredate: string;
  job_id: number;
  role_id: number;
  branch_id: number;
  company_id: number;
  superadmin: number;
  status: number;
  session_id: string;
  country_id: number;
  state_id: number;
  lga_id: number;
  bank_id: number;
  bank_account_no: string;
  employment_status: string;
  staff_category_id?: any;
  department_id: number;
  grade_id: number;
  line_manager_id: number;
  payroll_type: string;
  project_salary_category_id: number;
  last_login_at: string;
  last_login_ip: string;
  union_id: number;
  section_id: number;
  expat: number;
  non_payroll_provision_employee: number;
  confirmation_date: string;
  image_id?: any;
  last_promoted?: any;
  first_name: string;
  middle_name?: any;
  last_name: string;
  active: number;
  EDLEVEL?: any;
  account_num?: any;
  age?: any;
  bank_code?: any;
  bank_name?: any;
  confirmed?: any;
  dribble?: any;
  facebook?: any;
  grade?: any;
  instagram?: any;
  job_app_id?: any;
  job_reg_status?: any;
  kin_address?: any;
  kin_phonenum?: any;
  kin_relationship?: any;
  last_grade_change_date?: any;
  lga?: any;
  linemanager_id?: any;
  locale?: any;
  locked?: any;
  next_of_kin?: any;
  phone_num?: any;
  prev_grade?: any;
  sort_code?: any;
  state_origin_id?: any;
  twitter?: any;
  workdept_id?: any;
  religion?: any;
  probation_period?: any;
  created_at: string;
  updated_at: string;
  performance_category_id: number;
  bc_sync?: any;
  pali365_sync?: any;
  uses_pc: number;
  direct_salary_id?: any;
  years_of_service: number;
  months_of_service: number;
  user_image: string;
  my_status: string;
  job: Job;
}

interface Job {
  id: number;
  title: string;
  description?: any;
  department_id: number;
  qualification_id: number;
  parent_id?: any;
  personnel?: any;
  created_at: string;
  updated_at: string;
}